+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to Forge are documented in this file.
 
 ## [Unreleased]
 
+- Defaulted first-run sidebar appearance to glass mode.
+
 ## [0.0.14]